Rename UniqueField to UniqueProperty.
